xmonadrc: set env var for entire X session to prevent java doing the wrong thing
Most X window managers "reparent" new root-level windows, wrapping them
in an outer window that contains title bar, minimise button, etc.

xmonad does not do this. However, it's so expected at this point that the default
behaviour in many things is to assume it will happen and do weird things based on that.

We set an env var that tells java that this is NOT the case.