warn-on-camera-freeze: A hack to deal with a hardware issue
My camera occasionally freezes. Turning it off and back on fixes it, but I need to notice first.
This script scans dmesg for the error message that occurs when it happens, and makes a noise
that warns me it has happened.

It is intended to run permanently so it can keep warning for each instance.